feat: refactor queries

+++ src/utils/params.rs
@@ -2,6 +2,6 @@ use serde::Deserialize;
 
 #[derive(Deserialize, Debug)]
 pub struct DataApiQueryParams {
-    pub limit: Option<usize>,
+    pub limit: Option<i32>,
     pub ledger_index: Option<u32>,
 }
